Fixed two trivial problems related to the previous commit. Also, here's the information on how "special particles" work:
From dcr@astro.umd.edu Wed May 29 18:05:38 2002
Date: Wed, 27 Mar 2002 18:29:05 -0500 (EST)
From: Derek C. Richardson <dcr@astro.umd.edu>
Subject: J2/J4

Hey guys,

I've implemented a method in pkdgrav for handling "special" particles that
exert more than just a point-mass Newtonian force on other particles (and
themselves). Currently only oblateness effects (parameterized by
equatorial radius, J2, and J4) are supported, though there are hooks for a
GR "drag force" should Cole need it sometime in the future. I emphasized
generality and expandability in the coding so it should be easy to add
other effects or modify existing effects. Here's how it works:

The user supplies an extra text file that identifies which particles
(based on their *original* index, so for the moment only the COLLISIONS
version of pkdgrav is supported) are "special" and what makes them
special. For example, the following file describes a single special
particle, original index 0, for which oblateness terms (eq rad, J2, J4)
are supplied (this example is for Jupiter):

1
0 1 4.779e-4 14.75e-3 -580e-6

The first line is the number of special particles (currently limited to
just 3; you definitely don't want too many of these). Each subsequent line
describes a different particle. The first column is the original index
number (equivalent to the iOrder number if particle numbering doesn't
change). The second column indicates what special property the particle
has, in this case 1 = oblateness (it's a bit mask, so multiple properties
per particle are supported). The remaining columns are read in order of
increasing property number, in this case 3 columns for oblateness (eq rad,
J2, and J4).

This strategy means that the original data format (ss format in this case)
does not need to be changed. This is ideal when the number of special
particles is small. And it makes it easy to add new kinds of specialness.

Effects due to these special properties are applied immediately after
normal gravity has been computed (and before any external potentials have
been applied). First the code polls the processors for the current info on
all the special particles (for oblateness, we need the particle masses,
positions, and iOrder numbers). Then the special effects are applied to
all the particles. Only the processor tree is used, not the gravity tree.

The code recognizes two "super special" cases: 1) if the reference frame
is non-inertial (e.g. heliocentric, comet-centric, etc.), the net effect
of special particle forces are applied to the acceleration of the frame in
addition to all the other particles; 2) the reference frame itself can be
special (e.g. jovocentric with J2 & J4 terms) by specifying an original
index of -1 in the data file (only one such particle may be specified).

It compiles, and I think it works, but it still needs extensive testing.
Kevin should be able to help me with some of that, using SL9 sims.