This is a large change, the primary goal of which was to implement the Thacker and Couchman feedback as well as calculate star formation on user defined timesteps rather than just major timesteps.
----------------------------------------------------------------------
Modified Files:
	Makefile feedback.c main.c master.c outtype.c pkd.c pst.c
	smooth.c smoothfcn.c starform.c stiff.c supernova.c feedback.h
	master.h outtype.h parameters.h pkd.h pst.h smoothfcn.h
	starform.h stiff.h supernova.h
----------------------------------------------------------------------

The Thacker and Couchman feedback was introduced to us via James Wadsley,
who works with them at McMaster.  They describe the feedback in a 2000 paper.
James implemented the feedback in a separate definition called "SIMPLESF".

Tom liked the changes, so he suggested we integrate those changes into the
preexisting star formation.  By default the parameters are set to behave
exactly like the old star formation, but now there are 2 parameters that
you can set in the param file using yrs units (current popular values given):

dDeltaStarForm		= 1e6   # default = dDelta
dtCoolingShutoff	= 3e7	# default = 0.0

dDeltaStarForm is the parameter that sets the interval between star formation
and feedback events.
dtCoolingShutoff defines how many years cooling will be turned off for.  It
is the Thacker and Couchmann feedback.  Rob Thacker suggests that it was
Yepes 97 that came up with the original idea.

Several other star formation and feedback parameters were hard coded, and
now they can be set in the param file.  They are (with their default values):

dCStar = 0.1
dTempMax = 3e4
dSoftMin = 1.0
iStarFormRung = 0
dESN = 1e51

After we smooth the energy, mass, and cooling shutoff from the feedback,
we added an energy conservation check to the preexisting mass check.

There is still some stuff left to do:
1. change the number of particles that have their cooling shutoff.  Currently
all 32 nearest neighbors have their cooling shutoff for an amount of time
corresponding to the smoothing kernel.  This is highly resolution dependent
and we want to make the number of particles affected based on a more
constant value like distance from the forming star particle.

2. There is a problem in the masses that are used to calculate how much
mass is fedback by the wind feedback.  We need to make sure the calculations
are correct in all forms of feedback.

3.  The energy conservation check fails when the timesteps don't get smaller
than dDeltaStarform

4.  limit the number of particles that get their cooling calculated.  The
simulations currently run very slowly, and the major culprit is that cooling
is calculated for every gas particle on every minor timestep (not just
active particles).  We calculate the cooling for particles with their
cooling shutoff because if you just use the PdV energy, there is no guarantee
that your particle's temperature won't dip below 0, and it did and caused
crashes before we let cooling get calculated for every particle.

Getting back to other changes and fixes that included in this revision:

Anil implemented a wind feedback from massive stars.

Some brackets were missing meant that there was no feedback for a while.  Those
brackets were reinstated.

You can now safely star from checkpoint files even if you are not using a
director file to dump frames.

Removed all of the old SUPERNOVA definition code.  The new feedback replaces
most of its functionality.

The NFW profile in pkd.c was altered to what we wanted for our isolated
Milky Way test runs.

We added a new function construct to the smoothing architecture called
intTreeParticle.  It does stuff to the tree active particles over which
some quantity is going to be smoothed before the smoothing happens.  This
is much like initParticle except for the particles it affects.

stiff.c used code from Numerical Recipes that was not shared memory safe
and we replaced it with safe code.