BuildBinary(): removed alias that prevented optimization.
pkdGravAll(): Timing the individual stages of the gravity calculation impacts
the performance on some architectures.  Introduced a (compile time) variable
"bDetailTimer".  By default it is set to "0", so that only the entire loop is
timed, not "Walk", "Interact", and "Ewald" individually.  To get the
individual timings set it to non-zero.

fg(): changed exit() to mdlassert()

pkdDrift(): changed loop for better optimization on vector machines.