----------------------------------------------------------------------  Modified Files:  	main.c master.c outtype.c pkd.c pst.c smoothfcn.c starform.c  	master.h outtype.h parameters.h pkd.h pst.h smoothfcn.h  ----------------------------------------------------------------------
2 changes:

1.  Added bSNTurnOffCooling so that you can get old style feedback without
turning the cooling off.  The default for the parameter is 1, so cooling
turns off by default.  Set the parameter to 0 for old school feedback.

2.  A major rewrite of I/O.  A whole bunch of the cruft was taken out of
main.c and put into master.c or outtype.c.  Added a new iBinaryOutput = 6
that writes out snapshots in NChilada "SiX" (Simulation XML) format, which
writes binary output as separate files in a directory that contains gas, dark,
and star subdirectories.

The major noticeable change is that ASCII vector files are written
x
y
z
x
y
z

as opposed to the x's, then the y's, then the z's.  This was done so that
less swapping has to be done in parallel, which was slowing things down
hugely.  So, this should speed up output in parallel.

There are a couple of known issues with this change that will need to be fixed:
1.  Files are output in the SiX format that shouldn't be for various particle
types, so there might still be timeform files that get written for dark and
gas particles when it should only get written for stars.

2.  There are conflicts when the SiX format tries to create a directory
named the same thing as a tipsy output that for whatever reason was in
the directory already.  This could be fixed by adding ".data" to the end
of the filename for the director like it is done in Salsa.

3.  There is no way to read SiX format.