Starlog changes:
Star formation events are now recorded in a ".starlog" file.  See the README.starlog
for the format.  Also a "readstarlog.c" has been added to tipsy_tools to read this
file.  This file is updated by flushing an internal buffer of star formation events
whenever an output or a checkpoint is written.

This file replaces the (poorly functioning) *form files: rform, vform, rhoform
and Tform.  This also reduces the size of the particle structure and the
checkpoint files.  WARNING: checkpoint files using this version are incompatible
with previous versions.  MassForm and TimeForm files are still produced since the
physical state of the simulation depends on these quantities.  iGasOrder files are
also kept around only because the director "photogenic" files depend on them.

Modified Files:
 	main.c master.c master.h outtype.c outtype.h pkd.c pkd.h pst.c
 	pst.h starform.c starform.h
Added Files:
 	README.starlog